fix embarassing benchmark bug

+++ disassemble/disassemble.cc
@@ -1,8 +1,8 @@
 #include "unsq_eve/reduce.h"
 
 using T = char;
-using traits = unsq_eve::algorithm_traits<int, 256, 4>;
+using traits = unsq_eve::algorithm_traits<short, 256, 4>;
 
-int reduce(const char* f, const char* l) {
+short reduce(const char* f, const char* l) {
   return unsq_eve::reduce<traits>(f, l);
 }
